move get_organization_merged_id_by_query into pipeline.wikidata
