+++ mex/ff_projects/main.py
@@ -20,13 +20,13 @@ from mex.ff_projects.extract import (
     extract_ff_projects_organizations,
     extract_ff_projects_sources,
     filter_out_duplicate_source_ids,
-    get_organization_merged_id_by_query,
 )
 from mex.ff_projects.filter import filter_and_log_ff_projects_sources
 from mex.ff_projects.models.source import FFProjectsSource
 from mex.ff_projects.settings import FFProjectsSettings
 from mex.ff_projects.transform import transform_ff_projects_source_to_extracted_activity
 from mex.pipeline import asset, run_job_in_process
+from mex.pipeline.wikidata import get_organization_merged_id_by_query
 from mex.sinks import load
