+++ mex/odk/extract.py
@@ -3,11 +3,6 @@ from typing import Any
 
 from pandas import DataFrame, ExcelFile
 
-from mex.common.identity import get_provider
-from mex.common.models import ExtractedPrimarySource
-from mex.common.types import (
-    MergedOrganizationIdentifier,
-)
 from mex.common.wikidata.extract import search_organization_by_label
 from mex.common.wikidata.models.organization import WikidataOrganization
 from mex.odk.model import ODKData
@@ -94,33 +89,3 @@ def get_external_partner_and_publisher_by_label(
             external_partner_and_publisher_by_label[label] = organization
 
     return external_partner_and_publisher_by_label
-
-
-def get_organization_merged_id_by_query(
-    wikidata_organizations_by_query: dict[str, WikidataOrganization],
-    wikidata_primary_source: ExtractedPrimarySource,
-) -> dict[str, MergedOrganizationIdentifier]:
-    """Return a mapping from organizations to their stable target ID.
-
-    There may be multiple entries per unit mapping to the same stable target ID.
-
-    Args:
-        wikidata_organizations_by_query: Extracted organizations by query string
-        wikidata_primary_source: Primary source item for wikidata
-
-    Returns:
-        Dict with organization label and stable target ID
-    """
-    identity_provider = get_provider()
-    organization_stable_target_id_by_query = {}
-    for query, wikidata_org in wikidata_organizations_by_query.items():
-        identities = identity_provider.fetch(
-            had_primary_source=wikidata_primary_source.stableTargetId,
-            identifier_in_primary_source=wikidata_org.identifier,
-        )
-        if identities:
-            organization_stable_target_id_by_query[query] = (
-                MergedOrganizationIdentifier(identities[0].stableTargetId)
-            )
-
-    return organization_stable_target_id_by_query
