+++ mex/sumo/extract.py
@@ -4,12 +4,10 @@ from typing import Any
 import numpy as np
 from pandas import ExcelFile
 
-from mex.common.identity import get_provider
 from mex.common.ldap.connector import LDAPConnector
 from mex.common.ldap.models.actor import LDAPActor
 from mex.common.ldap.models.person import LDAPPersonWithQuery
 from mex.common.ldap.transform import analyse_person_string
-from mex.common.models import ExtractedPrimarySource
 from mex.common.wikidata.extract import search_organization_by_label
 from mex.common.wikidata.models.organization import WikidataOrganization
 from mex.sumo.models.cc1_data_model_nokeda import Cc1DataModelNoKeda
@@ -205,33 +203,3 @@ def extract_sumo_organizations(
         if label and (org := search_organization_by_label(label)):
             sumo_resource_organizations[label] = org
     return sumo_resource_organizations
-
-
-def get_organization_merged_id_by_query(
-    sumo_organizations: dict[str, WikidataOrganization],
-    wikidata_primary_source: ExtractedPrimarySource,
-) -> dict[str, str]:
-    """Return a mapping from organizations to their stable target ID.
-
-    There may be multiple entries per unit mapping to the same stable target ID.
-
-    Args:
-        sumo_organizations: Iterable of extracted organizations
-        wikidata_primary_source: Primary source item for wikidata
-
-    Returns:
-        Dict with organization label and stable target ID
-    """
-    identity_provider = get_provider()
-    organization_stable_target_id_by_query = {}
-    for query, wikidata_organization in sumo_organizations.items():
-        identities = identity_provider.fetch(
-            had_primary_source=wikidata_primary_source.stableTargetId,
-            identifier_in_primary_source=wikidata_organization.identifier,
-        )
-        if identities:
-            organization_stable_target_id_by_query[query] = str(
-                identities[0].stableTargetId
-            )
-
-    return organization_stable_target_id_by_query
